Add measured-latency gate: decide on real EXPLAIN ANALYZE, not just hypopg
Upgrades the evolution loop from a single estimated-cost gate to a two-tier
gate, so changes are kept based on REAL measured latency rather than the
planner's estimate (which is only loosely correlated with wall-clock time):

- Tier 1 (unchanged): hypopg estimated cost drives the evolutionary search and
  a cheap policy pre-filter — no production impact.
- Tier 2 (new): the winning candidate is validated by a measured trial —
  EXPLAIN (ANALYZE) latency across the weighted workload — and kept only if the
  measurement agrees. Runs on a shadow replica when `shadow_database_url` is set
  (zero production impact: build → measure → drop on the replica, apply to the
  primary only if it passes), otherwise an in-place trial with guaranteed
  auto-rollback.

Reliability of the measurement (the hard part):
- best-of-N per query (the fastest run is the least-noisy estimate of true cost)
- a noise floor: sub-millisecond queries can't veto, and a regression must
  exceed both a relative % and an absolute ms margin
- a realistic in-place regression tolerance (an index build perturbs the buffer
  cache), tightenable via a quiet shadow replica

Provenance now records predicted (hypopg) vs measured for every change, so the
estimate-vs-reality gap is visible in pistol.evolution_history.

Also fixes a latent demo-seed bug: intervals built by string-concatenating
random()*N could emit scientific notation (e.g. "6.7e-05 days") that ::interval
rejects; now cast to integer days.

New module src/measure.rs; apply.rs refactored to online build/drop primitives
that the engine composes into the trial. Unit tests cover the gate math (weighted
summary, noise-floor exclusion); the integration test now validates a measured
apply and a measured auto-rollback. Docs updated.

+++ src/demo.rs
@@ -23,7 +23,7 @@ pub async fn seed(pool: &PgPool) -> anyhow::Result<()> {
 
     INSERT INTO public.students
       SELECT g, 1+(random()*49)::int, 'Student '||g, 1+(random()*11)::int,
-             now() - (random()*400||' days')::interval
+             now() - ((random()*400)::int||' days')::interval
       FROM generate_series(1,5000) g;
 
     INSERT INTO public.classes
@@ -34,18 +34,18 @@ pub async fn seed(pool: &PgPool) -> anyhow::Result<()> {
 
     INSERT INTO public.enrollments
       SELECT g, 1+(random()*4999)::int, 1+(random()*499)::int,
-             now() - (random()*300||' days')::interval,
+             now() - ((random()*300)::int||' days')::interval,
              (ARRAY['active','active','active','dropped','completed'])[1+(g%5)]
       FROM generate_series(1,30000) g;
 
     INSERT INTO public.assignments
       SELECT g, 1+(random()*499)::int, 'Assignment '||g,
-             now() + (random()*30||' days')::interval
+             now() + ((random()*30)::int||' days')::interval
       FROM generate_series(1,5000) g;
 
     INSERT INTO public.submissions
       SELECT g, 1+(random()*4999)::int, 1+(random()*4999)::int,
-             now() - (random()*200||' days')::interval,
+             now() - ((random()*200)::int||' days')::interval,
              (random() < 0.6), (random()*100)::int
       FROM generate_series(1,100000) g;
 
@@ -53,13 +53,13 @@ pub async fn seed(pool: &PgPool) -> anyhow::Result<()> {
       SELECT g, 1+(random()*4999)::int, 1+(random()*499)::int, 1+(random()*49)::int,
              (ARRAY['in_progress','completed','not_started'])[1+(g%3)],
              (random()*100)::int,
-             now() - (random()*365||' days')::interval
+             now() - ((random()*365)::int||' days')::interval
       FROM generate_series(1,150000) g;
 
     INSERT INTO public.activity_events
       SELECT g, 1+(random()*4999)::int,
              (ARRAY['login','view','submit','comment'])[1+(g%4)],
-             now() - (random()*365||' days')::interval
+             now() - ((random()*365)::int||' days')::interval
       FROM generate_series(1,200000) g;
 
     ANALYZE;
